renames connected app methods
